authorization added

+++ controller/expense.js
@@ -6,7 +6,7 @@ const addexpense =  (req,res)=>{
     if(expenseamount==undefined|| expenseamount.length===0){
         return res.status(400).json({success:false,message:'parameters missing'})
     }
-    Expense.create({expenseamount,description,category}).then(expense=>{
+    Expense.create({expenseamount,description,category,userId:req.user.id}).then(expense=>{
         return res.status(201).json({expense,success:true});
     }).catch(err=>{
         return res.status(500).json({success:false,error: err})
@@ -14,7 +14,7 @@ const addexpense =  (req,res)=>{
 }
 
 const getexpenses  =  (req,res)=>{
-    Expense.findAll().then(expenses=>{
+    Expense.findAll({where:{userId: req.user.id }}).then(expenses=>{
         return res.status(200).json({expenses,success:true})
 
     }).catch(err=>{
@@ -25,7 +25,10 @@ const getexpenses  =  (req,res)=>{
 
 const deleteexpense =  (req,res)=>{
     const expenseid  =  req.params.expenseid
-    Expense.destroy({where:{id:expenseid}}).then(()=>{
+    Expense.destroy({where:{id:expenseid,userId:req.user.id}}).then((noofrows)=>{
+        if(noofrows===0){
+            return res.status(404).json({success:false,message:'Expense doesnot belong to the user'})
+        }
         return res.status(200).json({success:true,message:"deleted successfully"})
     }).catch(err=>{
         console.log(err);
